Fix bug where only one user could be created
Only one user could be created on the site,
because the username field was a blank string by default
but unique in the database.
Now its null by default,
which is not considered when checking for uniqueness.

We are not actually using the username field,
but the email field for login.
But the username field is left in for legacy users.

Added comments for clarification.